+++ CHANGES.md
@@ -4,6 +4,7 @@
 
  * Added option to display table-of-contents on namespace pages
  * Improve formatting of large objects when used as a card body
+ * Add `resettable` option, which displays a button to remount the card body
 
 ## v0.6.2